+++ attributes/default.rb
@@ -135,6 +135,11 @@ default['openstack']['orchestration']['api']['auth']['insecure'] = false
 # Keystone role for heat template-defined users. (string value)
 default['openstack']['orchestration']['heat_stack_user_role'] = nil
 
+# Keystone domain id which contains heat template-defined users.
+# If this option is set, stack_user_domain_name option
+# will be ignored. (string value)
+default['openstack']['orchestration']['stack_user_domain_id'] = nil
+
 # Keystone domain name which contains heat template-defined users. (string value)
 default['openstack']['orchestration']['stack_user_domain_name'] = nil